Refactor sessions screen state into hooks
